Issue #72 Fix bug with determining whether to add next_offset.
When offset is 0 and limit is 1, the next_offset should only be shown if the row count is greater than the next offset (i.e. is 2 or greater).
This is due to offset starting from 0 and count starting from 1 (i.e. a count of 1 satisfies/maps to the offset of 0)