fix: Update analytics recording to record every module download.
If environments are not enabled, or a single environment-less key has been provided, use a 'default' environment name for recording module.
If environments are enabled and key for a given environment has been provided, record download against this environment.
For all other downloads, record the download using an empty environment, which will be used for download counts

Issue #119