db: Fix method of storing/returning transaction to store connection.
Queries are executed on the connection that has performed the begin, rather than the object returned from the begin (or the object returned from the enter of the begin.
The parent connection is now stored in context, which is then returned when in the context of a transaction.
Create a fake wrapper than returns the transaction connection, which can then used in a 'with-statment' (similar to how a connection is handled outside of a transaction).

Issue #101