db: Remove foreign key constraint from analytics table for parent module version ID.
This allows the analaytics rows to be left 'dangling' whilst re-creating new rows for replaced module versions.
This would be an ideal candidate for a 'deferred' foreign key, in-which the integrity is checked at the end of a transaction - but MySQL does not support this.
Alternative solutions maybe to re-use the same module version row, but this would involve resetting all columns, somehow, which isn't available in the architecture (maybe using an ORM would solve this).
Since the table is used either to obtain global counts or per module version counts, the 'previous' foreign key was not used to lookup a module version from a given analytics row - this would need to be carefully used if implemented in the future.

Issue #153