fix: Fix get_all method of Namespace class to return all namespaces when only_published is False.
The previous query joined to the module verison table with the latest_version column.
This column is empty when the module provider does not have a published non-beta version, meaning that namespaces with only a module with a beta/non-published version would not be included.
When only_published is False, now perform basic select from module_provider table, selecting/grouping/ordering by the namespace column

Issue #167