fix: Correctly return protocol in all integration URLs
Update integrations function/endpoint (and initial setup endpoint) to return only the path of the URL for the integration.
Add javascript helper to convert the path to a full URL, including the current protocol and domain (and port, if non-standard).
Use this new method for initial setup page and the integrations tab of the module provider page.
Update tests to match the new expected http endpoints (since tests run over http)

Issue #179