fix: Use database transaction when performing module provider creation in server endpoint.
Update Transaction object to return instance of transaction class, rather than connection.
Property properties for DB transaction object and connection object.
Rollback transaction when an error occurs during module provider creation - this is done as any exceptions are caught and gracefully returned to the user, so the transaction is not automatically rolled back

Issue #191