fix: Support examples obtaing example cost for modules that use other modules hosted in Terrareg.
Add configuration to hold 'public' hostname of the registry.
Generate internal auth token and pass hostname of registry to infracost, so that it can authenticate to the registry.
This is required to bypass any checks for analytics token in the URL and to stop the download from being recorded.
Move checking of valid download to record_module_version_download, which check if token is present, required and whether the auth token is the new infracost internal token

Issue #226