Update module provider version dropdown/warning to always process beta/unpublished warning.
Previously, beta and unpublished warning was only shown when selected version was not present in the version dropdown (due to only non-beta published versions being shown in the dropdown).
Now, all versions can be present in the list, so the code to display the warnings was not executed.

Update warnings so that only one is displayed in the order of: unpublished, beta, non-latest.
Update version in drop down to shown both beta and unpublished attributes, if applicable, rather than prefering the beta flag and showing only one

Issue #161