feat: Retain publish state when re-indexing module version
Update prepare_module and _create_db_row for module version to return whether original version was published, if it existed.
Update APIs that create module versions to publish module version after index, if a previous version existed that was published.
Update tests for APIs for creating module versions to test this functionality.

Issue #299