feat: Add support for external images in README and custom markdown tabs.
Add 'img' to allowed tags in HTML sanitisation and 'src' attribute of 'img' to allowed attributes.
Update markdown parser to strip the 'src' attribute for any relative paths, as the path will be relative to the original repository and will not work in Terrareg.
Add tests to check relative and full URL src attributes of images

Issue #370