fix: Attempt to find find terraform files with backend configuration and create override file during extraction.
The previous method of disabling the backend during init did not work - the init worked successfully, however the graph command fails, requiring terraform to be initialised with the backend.
Replace with method that scans all terraform files in the directory for backend configurations and creates and override file with the same name that replaces the backend configuration with a local state.
KNOWN ISSUE: This will not help in situations where the state is configured in a submodule (either elsewhere in the repository or in a remote module). However, this is not a regression, as this previously would have also not worked.

Issue #395