fix: Fix erronious links in converted README markdown.
If an HTML element is found, which is not processed with a custom anchor, the starttag is not reset from
the previously processed element.
This means after processing a header, which is converted to an anchor link, subsequent unprocessable tags (such as line breaks) were replaced with the same link tag as the previously processed header.

Issue #402