fix: Fix bug where CWD of main process is changing during module extraction.
Replace use of shutil.make_archive, as this function is not thread safe and changes the CWD of the main process during archive creation.
The CWD is used by configuration to determine the data directory for creating archives.
This caused errors when extracting multiple modules, as the CWD would change to be a temporary directory that would have caused modules to be stored in temporary directory and was also removed after the previous extraction was complete (race conditions also caused failures due to CWD no longer existing)

Add installation of 'zip' package required to perform zipping of module

Issue #404