feat: Redirect user to login when authentication is required.
Remove authentication requirement for config endpoint.
Update is_authenticated endpoint to return without authentication and to specify if read access is allowed.
Update the template setupPageLogin to check if user is authenticated using attribute of the response from is_authenticated and redirect to login page if authentication is required

Issue #289